spark/requirements.txt
request
joblib
redis
scikit-learn
pandas